[+] before armagedon
