+++ src/components/footer/Footer.css
@@ -1,0 +1,28 @@
+.footer {
+   margin-left: 0;
+   margin-right: 0;
+   margin-top: 4%;
+   background-color:  #ffb515;
+}
+
+.footer_details {
+   margin-top: 5px;
+   font-size: 0.9rem;
+   color: white;
+}
+
+.redes{
+   margin-top: 1%;
+   padding-bottom: 1%;
+}
+.icono{
+   margin-left: 1%;
+   margin-right: 1%;
+}
+.copyright p{
+   margin-bottom: 5px;
+}
+.contacto p{
+   margin-bottom: 0;
+   padding-bottom: 0;
+}
