--- src/components/index/Index.css
@@ -1,176 +1,0 @@
-.fondo img{
-   z-index: -1;
-   margin-top: -25vh;
-   text-align: center;  
-   height: 85vh;
-   width: 100%;
-   object-fit: cover;
-}
-
-.typing {
-   z-index: 1;
-   position: relative;
-   height: 200px;
-}
-.presentacion.fondo h1{
-   padding-top: 35vh;
-   color: rgb(245, 247, 244);
-   font-size: 60px;
-   font-weight: 400;
-   text-shadow: rgb(0, 0, 0) 1px 0 1px;
-}
-
-.bienvenido h2{
-   text-align: center;
-   color:rgb(26, 85, 2);
-   font-weight: 400;
-   margin-top: 0;
-}
-
-.laredsocial h2{
-   margin-top: 40px;
-   font-weight: 500;
-}
-
-.nosotros {
-   margin-top: 20px;
-   margin-left: 0;
-   margin-right: 0;
-   text-align: center;
-}
-.nosotros_texto {
-   font-size: 1rem;
-   text-align: justify;
-   margin-top: 15px;
-}
-.nosotros_texto h1{
-
-   text-align: center;
-   color:rgb(26, 85, 2);
-   font-size: x-large;
-   padding-bottom: 0.8em;
-   border-bottom: solid 1px #ffb515;
-   margin-bottom: 20px;
-}
-.nosotros_img {
-   margin-top: 50px;
-   margin-bottom: 50px;
-   text-align: center;
-}
-.nosotros_img img{
-   width: 10%;
-   text-align: center;
-   margin-top: 0;
-}
-
-.footer {
-   margin-left: 0;
-   margin-right: 0;
-   margin-top: 4%;
-   background-color:  #ffb515;
-}
-
-.footer_details {
-   margin-top: 5px;
-   font-size: 0.9rem;
-   color: white;
-}
-
-.video {
-   margin-top: 100px;
-}
-
-.redes{
-   margin-top: 1%;
-   padding-bottom: 1%;
-}
-.icono{
-   margin-left: 1%;
-   margin-right: 1%;
-}
-.copyright p{
-   margin-bottom: 5px;
-}
-.contacto p{
-   margin-bottom: 0;
-   padding-bottom: 0;
-}
-
-@media (max-width: 600px) {
-   .fondo img{
-      height: 80vh;
-   }
-
-   .presentacion.fondo h1{
-      font-size: 40px;
-      font-weight: 400;
-      padding-top: 35vh;
-   }
-   .bienvenido {
-      text-align: center;
-   }
-   .bienvenido h2 {
-      text-align: center;
-      font-size: 1.3rem;
-      font-weight: 400;
-      color:rgb(26, 85, 2);
-   }
-   .laredsocial h2{
-      font-size: 1.5rem;
-      margin-top: 20px;
-      margin-bottom: 40px;
-   }
-   
-   .embedResponsive{
-      text-align: center;
-   }
-   .video {
-      margin-top: 30px;
-      margin-bottom: 20px;
-   }
-
-   .nosotros_texto {
-      margin: 50px 20px 0px 20px;
-    
-   }
-   .nosotros_texto h1{
-      text-align: center;
-      font-size: 20px;     
-      color:rgb(26, 85, 2);
-   }
-   .footer_details {
-      margin-top: 5px;
-      margin-bottom: 5px;
-      font-size: 10px;
-      color: white;
-      text-align: center;
-   }
-   .nosotros_img {
-      margin-top: 0px;
-      margin-bottom: 0px;
-   }
-   .nosotros_img img{
-      margin-top: 10px;
-      margin-bottom: 10px;
-      width: 20%;
-   }
-}
-
-
-@media (min-width: 1300px)   {
-   .fondo img {
-      margin-top: -20vh;
-   }
-   .nosotros_img img{
-      width: 5%;
-      text-align: center;
-      margin-top: 0;
-   }
-   .bienvenido h2 {
-      margin-top: 50px;
-   }
-   .nosotros_texto {
-      font-size: 1.15rem;
-   }
-   
-}
