+++ src/components/index/info/Info.css
@@ -1,0 +1,103 @@
+.bienvenido h2{
+   text-align: center;
+   color:rgb(26, 85, 2);
+   font-weight: 400;
+   margin-top: 0;
+}
+
+.video {
+   margin-top: 100px;
+   border-radius:20px;
+}
+
+.laredsocial h2{
+   margin-top: 40px;
+   font-weight: 500;
+}
+
+.nosotros {
+   margin-top: 20px;
+   margin-left: 0;
+   margin-right: 0;
+   text-align: center;
+}
+.nosotros_texto {
+   font-size: 1rem;
+   text-align: justify;
+   margin-top: 15px;
+}
+.nosotros_texto h1{
+
+   text-align: center;
+   color:rgb(26, 85, 2);
+   font-size: x-large;
+   padding-bottom: 0.8em;
+   border-bottom: solid 1px #ffb515;
+   margin-bottom: 20px;
+}
+.nosotros_img {
+   margin-top: 50px;
+   margin-bottom: 50px;
+   text-align: center;
+}
+.nosotros_img img{
+   width: 10%;
+   text-align: center;
+   margin-top: 0;
+}
+
+@media (max-width: 600px) {
+    .video {
+      margin-top: 30px;
+      margin-bottom: 20px;
+    }
+    .bienvenido {
+        text-align: center;
+    }
+    .bienvenido h2 {
+        text-align: center;
+        font-size: 1.3rem;
+        font-weight: 400;
+        color:rgb(26, 85, 2);
+    }
+   .laredsocial h2{
+      font-size: 1.5rem;
+      margin-top: 20px;
+      margin-bottom: 40px;
+   }
+
+   .nosotros_texto {
+      margin: 50px 20px 0px 20px;
+    
+   }
+
+   .nosotros_texto h1{
+      text-align: center;
+      font-size: 20px;     
+      color:rgb(26, 85, 2);
+   }
+   .nosotros_img {
+      margin-top: 0px;
+      margin-bottom: 0px;
+   }
+   .nosotros_img img{
+      margin-top: 10px;
+      margin-bottom: 10px;
+      width: 20%;
+   }
+}
+
+@media (min-width: 1300px)   {
+   .nosotros_img img{
+      width: 5%;
+      text-align: center;
+      margin-top: 0;
+   }
+   .bienvenido h2 {
+      margin-top: 50px;
+   }
+   .nosotros_texto {
+      font-size: 1.15rem;
+   }
+   
+}
