+++ src/components/navbar/Navbar.css
@@ -3,7 +3,7 @@
     display: flex;
     justify-content: space-between;
     background-color: white;
-    width: 100%;
+    width: 100vw;
     z-index: 2;
     margin-top: 0;
     margin-bottom: 0px;
